% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dief.R
\name{plotAnswerTrace}
\alias{plotAnswerTrace}
\title{Plot the answer trace of approaches}
\usage{
plotAnswerTrace(inputtrace, inputtest)
}
\arguments{
\item{inputtrace}{dataframe with the answer trace. Attributes of the dataframe: test, approach, answer, time.}

\item{inputtest}{string that specifies the specific test to analyze from the answer trace.}
}
\description{
This function plots the answer trace of the approaches when executing a given test.
}
\examples{
 
plotAnswerTrace(traces, "Q9.sparql")
}
\seealso{
diefk, dieft
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
