% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment1}
\alias{plotExperiment1}
\title{plotExperiment1}
\usage{
plotExperiment1(allmetrics)
}
\description{
This function plots the results reported in Experiment 1.
}
\seealso{
experiment1, diefk2
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
