% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment1Query}
\alias{plotExperiment1Query}
\title{plotExperiment1Query}
\usage{
plotExperiment1Query(allmetrics, q)
}
\arguments{
\item{allmetrics}{dataframe with the results of all the metrics in Experiment 1.}

\item{q}{the id of the selected query to plot.}
}
\description{
This function plots the results reported in Experiment 1 for a single given query.
}
\seealso{
experiment1, plotExperiment1
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{dieft,}
