% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment1Test}
\alias{plotExperiment1Test}
\title{plotExperiment1Test}
\usage{
plotExperiment1Test(allmetrics, q)
}
\arguments{
\item{allmetrics}{dataframe with the results of all the metrics in Experiment 1.}

\item{q}{the id of the selected test to plot.}
}
\description{
This function plots the results reported for a single given test in "Experiment 1" in in Acosta, M., Vidal, M. E., & Sure-Vetter, Y. (2017). Diefficiency metrics: Measuring the continuous efficiency of query processing approaches. In International Semantic Web Conference (pp. 3-19). Springer, Cham. <doi:10.1007/978-3-319-68204-4_1>..
}
\seealso{
experiment1, plotExperiment1
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{dieft,}
