% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment2}
\alias{plotExperiment2}
\title{plotExperiment2}
\usage{
plotExperiment2(diefkDF)
}
\arguments{
\item{diefkDF}{dataframe with the results of Experiment 2.}
}
\description{
This function plots the results reported in Experiment 2.
}
\seealso{
experiment2, diefk2
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
