% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment2Query}
\alias{plotExperiment2Query}
\title{plotExperiment2Query}
\usage{
plotExperiment2Query(diefkDF, q)
}
\arguments{
\item{diefkDF}{dataframe resulting from Experiment 2.}

\item{q}{the id of the selected query to plot.}
}
\description{
This function plots the results reported in Experiment 2 for a single given query.
}
\seealso{
experiment2, plotExperiment2
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
