% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment2Test}
\alias{plotExperiment2Test}
\title{Generate radar plots that compare dief@k at different answer completeness in a specific test as in  <doi:10.1007/978-3-319-68204-4_1>}
\usage{
plotExperiment2Test(diefkDF, q)
}
\arguments{
\item{diefkDF}{dataframe resulting from Experiment 2.}

\item{q}{id of the selected test to plot.}
}
\description{
This function plots the results reported for a single given test in "Experiment 2" in Acosta, M., Vidal, M. E., & Sure-Vetter, Y. (2017) <doi:10.1007/978-3-319-68204-4_1>.
"Experiment 2" measures the continuous efficiency of approaches when producing the first 25%, 50%, 75%, and 100% of the answers.
}
\examples{
results2 <- experiment2(traces)
plotExperiment2Test(results2, "Q9.sparql")
}
\seealso{
experiment2, plotExperiment2
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
