% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dief.R
\name{diefk}
\alias{diefk}
\title{diefk}
\usage{
diefk(inputtrace, inputquery, k = -1)
}
\arguments{
\item{inputtrace}{Dataframe with the answer trace. Attributes of the dataframe: query, approach, tuple, time.}

\item{inputquery}{String that specifies the specific query to analyze from the answer trace.}

\item{k}{to compute diefk. By default, the function computes the minimum of the total number of answers produced by the approaches.}
}
\description{
This function computes the dief@k metric at a given k (number of answers).
}
\seealso{
dieft, diefk2, plotAnswerTrace
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
