% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dief.R
\name{diefk2}
\alias{diefk2}
\title{diefk2}
\usage{
diefk2(inputtrace, inputquery, kp = -1)
}
\arguments{
\item{inputtrace}{Dataframe with the answer trace. Attributes of the dataframe: query, approach, tuple, time.}

\item{inputquery}{String that specifies the specific query to analyze from the answer trace.}

\item{kp}{to compute diefk. By default and when kp=1.0, this function behaves the same as diefk. It computes the kp portion of of minimum of of number of answers  produced by the approaches.}
}
\description{
This function computes the dief@k metric at a given kp (percentage of answers).
}
\seealso{
dieft, diefk, plotAnswerTrace
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
