% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dief.R
\name{diefk2}
\alias{diefk2}
\title{diefk2}
\usage{
diefk2(inputtrace, inputtest, kp = -1)
}
\arguments{
\item{inputtrace}{Dataframe with the answer trace. Attributes of the dataframe: test, approach, answer, time.}

\item{inputtest}{String that specifies the specific test to analyze from the answer trace.}

\item{kp}{Portion of answers to compute diefk (between 0.0 and 1.0). By default and when kp=1.0, this function behaves the same as diefk. It computes the kp portion of of minimum of of number of answers  produced by the approaches.}
}
\description{
This function computes the dief@k metric at a given kp (percentage of answers).
}
\seealso{
dieft, diefk, plotAnswerTrace
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
