% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dief.R
\name{dieft}
\alias{dieft}
\title{dieft}
\usage{
dieft(inputtrace, inputquery, t = -1)
}
\arguments{
\item{inputtrace}{Dataframe with the answer trace. Attributes of the dataframe: query, approach, tuple, time.}

\item{inputquery}{String that specifies the specific query to analyze from the answer trace.}

\item{t}{to compute dieft. By default, the function computes the minimum of the execution time among the approaches in the answer trace.}
}
\description{
This function computes the dief@t metric.
}
\seealso{
diefk, diefk2, plotAnswerTrace
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{dieft,}
