% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{experiment1}
\alias{experiment1}
\title{Compares dief@t with other benchmark metrics as in <doi:10.1007/978-3-319-68204-4_1>}
\usage{
experiment1(traces, metrics)
}
\arguments{
\item{traces}{dataframe with the result of the traces. The structure of this dataframe is as follows: "test,approach,tuple,time".}

\item{metrics}{dataframe with the result of the other metrics. The structure of this dataframe is as follows: "test,approach,tfft,totaltime,comp".}
}
\description{
This function repeats the results reported in "Experiment 1" in Acosta, M., Vidal, M. E., & Sure-Vetter, Y. (2017) <doi:10.1007/978-3-319-68204-4_1>.
Experiment 1 compares the performance of querying approaches when using metrics defined in the literature (total execution time, time for the first tuple, throughput, and completeness) and the metric dieft@t.
}
\examples{
# To fully reproduce the experiments download the full files and load them using read.csv:
# traces is available at <https://figshare.com/files/9625852>
# metrics is available at <https://figshare.com/files/9660316>
results1 <- experiment1(traces, metrics)
}
\seealso{
experiment2, dieft
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{dieft,}
