% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{experiment1}
\alias{experiment1}
\title{experiment1}
\usage{
experiment1()
}
\description{
This function reproduces the results reported in Experiment 1.
}
\seealso{
experiment2, dieft
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{dieft,}
