% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{experiment1}
\alias{experiment1}
\title{experiment1}
\usage{
experiment1(traces_file, metrics_file)
}
\arguments{
\item{traces_file}{CSV file with the result of the traces. The structure of this file is as follows: "test,approach,tuple,time".}

\item{metrics_file}{CSV file with the result of the other metrics. The structure of this file is as follows: "test,approach,tfft,totaltime,comp".}
}
\description{
This function reproduces the results reported in Experiment 1.
}
\seealso{
experiment2, dieft
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{dieft,}
