% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{experiment2}
\alias{experiment2}
\title{experiment2}
\usage{
experiment2(traces_file)
}
\arguments{
\item{traces_file}{CSV file with the result of the traces. The structure of this file is as follows: "test,approach,tuple,time".}
}
\description{
This function reproduces the results reported in Experiment 2.
}
\seealso{
experiment1, diefk2
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
