% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dief.R
\name{plotAnswerTrace}
\alias{plotAnswerTrace}
\title{plotAnswerTrace}
\usage{
plotAnswerTrace(inputtrace, inputquery)
}
\arguments{
\item{inputtrace}{Dataframe with the answer trace. Attributes of the dataframe: query, approach, tuple, time.}

\item{inputquery}{String that specifies the specific query to analyze from the answer trace.}
}
\description{
This function plots the answer trace of a given query.
}
\seealso{
diefk, dieft
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
