% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dief.R
\name{plotAnswerTrace}
\alias{plotAnswerTrace}
\title{plotAnswerTrace}
\usage{
plotAnswerTrace(inputtrace, inputtest)
}
\arguments{
\item{inputtrace}{Dataframe with the answer trace. Attributes of the dataframe: test, approach, answer, time.}

\item{inputtest}{String that specifies the specific test to analyze from the answer trace.}
}
\description{
This function plots the answer trace of a given test.
}
\seealso{
diefk, dieft
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
