% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment2Test}
\alias{plotExperiment2Test}
\title{plotExperiment2Test}
\usage{
plotExperiment2Test(diefkDF, q)
}
\arguments{
\item{diefkDF}{dataframe resulting from Experiment 2.}

\item{q}{the id of the selected test to plot.}
}
\description{
This function plots the results reported for a single given test in "Experiment 2" in Acosta, M., Vidal, M. E., & Sure-Vetter, Y. (2017). Diefficiency metrics: Measuring the continuous efficiency of query processing approaches. In International Semantic Web Conference (pp. 3-19). Springer, Cham. <doi:10.1007/978-3-319-68204-4_1>..
}
\seealso{
experiment2, plotExperiment2
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
