% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduceExperiments.R
\name{plotExperiment2Test}
\alias{plotExperiment2Test}
\title{plotExperiment2Test}
\usage{
plotExperiment2Test(diefkDF, q)
}
\arguments{
\item{diefkDF}{dataframe resulting from Experiment 2.}

\item{q}{the id of the selected test to plot.}
}
\description{
This function plots the results reported in Experiment 2 for a single given test.
}
\seealso{
experiment2, plotExperiment2
}
\author{
Maribel Acosta
}
\keyword{diefficiency}
\keyword{diefk,}
